Allow using plain lyrics (#4)
* Allow using plain lyrics

Allow the option to use plain lyrics if synced aren't available. They
can be fed into a LRC generator for making proper synced lyrics, and for
some people plain lyrics is better than no lyrics.

Changes the CLI parser to `clap` to make it easier to declare flags and
parse the ways someone might pass them. Keeps the existing behaviour as
default (`lyricsrs /path/to/music/dir` works and skips plain lyrics) and
adds the `--allow-plain` CLI flag to tell `lyricsrs` to write plain
lyrics if that's all that's available.

* Update src/main.rs

Co-authored-by: td <td@technodisaster.com>

* Update main.rs

---------

Co-authored-by: td <td@technodisaster.com>